+++ ARBFN/interchange.h
@@ -101,17 +101,26 @@ struct FFieldNodeData {
 };
 
 /**
- * @brief Interchange, but for ffield fixes. This only happens once, upon simulation initialization.
+ * @brief Interchange, but for ffield fixes. This may only happen once
+ * (upon simulation initialization), or may be reoccurring every once in a while. In
+ * the latter case, the final two arguments will be used to "dump" atom data to the
+ * controller before refreshing.
  * @param _start A 3-tuple (x, y, z) of the lowest corner of the simulation box.
  * @param _bin_widths A 3-tuple for the x, y, and z spacing of the nodes.
  * @param _node_counts The number of nodes per side. A 3-tuple of the x, y, and z.
  * @param _controller_rank The rank of the controller within the provided communicator
  * @param _comm The MPI communicator to use
+ * @param _atoms_to_send_size (optional) If provided, the number of atoms in
+ * `_atoms_to_send`. If 0, don't send any atoms.
+ * @param _atoms_to_send (optional) If the size is positive, send these to the
+ * controller along with the request.
  * @returns A std::list of the data to be added
  */
 std::list<FFieldNodeData> ffield_interchange(const double _start[3], const double _bin_widths[3],
-                                             const unsigned int _bin_counts[3],
-                                             const unsigned int &_controller_rank, MPI_Comm &_comm);
+                                             const unsigned int _node_counts[3],
+                                             const unsigned int &_controller_rank, MPI_Comm &_comm,
+                                             const unsigned int &_atoms_to_send_size = 0,
+                                             const AtomData _atoms_to_send[] = {});
 
 /**
  * @brief Send the given atom data, then receive the given fix data. This is blocking, but does not allow worker-side gridlocks.
@@ -124,8 +133,7 @@ std::list<FFieldNodeData> ffield_interchange(const double _start[3], const doubl
  * @returns true on success, false on failure
  */
 bool interchange(const size_t &_n, const AtomData _from[], FixData _into[], const double &_max_ms,
-                 const unsigned int &_controller_rank, MPI_Comm &_comm,
-                 const bool &_expect_response = true);
+                 const unsigned int &_controller_rank, MPI_Comm &_comm);
 
 /**
  * @brief Sends a registration packet to the controller.